refactor: extract useLayoutPanels hook from App.tsx — reduce god component churn (#195)
